+++ test/fixtures/react-router-search-params-invalid/src/pages/index.tsx
@@ -1,0 +1,6 @@
+import { useSearchParams } from "react-router-dom"
+
+export default function Page() {
+  const [searchParams, setSearchParams] = useSearchParams()
+  return <button onClick={() => setSearchParams(searchParams)}>Update</button>
+}